Fix missing <script>s on imported posts
For some reason the scripts I used to import my WordPress posts failed
to import any javascript into the markdown files, but they *did* import
with one of the tools I used (I used several), which imported to html
files.

I wrote a script `utils/fix_js.sh` that used `ag` to search for
`<script.*</script>` in the `.html` files and used that filelist to open
a vim split window for each file and the location in the quickfix using
`vimgrep`. Was still a little tedious, but much better than it might
have been.